completed state
