+++ src/components/RevealAnswer.tsx
@@ -2,5 +2,14 @@ import React, { useState } from "react";
 import { Button } from "react-bootstrap";
 
 export function RevealAnswer(): JSX.Element {
-    return <div>Reveal Answer</div>;
+    
+    const [visible, setVisible] = useState<boolean>(false);
+
+    function flipVisibility(): void {
+        setVisible(!visible);
+    }
+    
+    return <div> 
+        <div>Reveal Answer</div><Button onClick={flipVisibility}>Reveal Answer</Button> {visible && <div>42</div>}
+    </div>;
 }
